pyproject.toml
[project]
name = "langgraph-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.20",
    "langchain-mcp-adapters>=0.1.10",
    "langgraph>=0.6.7",
    "langgraph-api>=0.4.22",
    "langgraph-cli>=0.4.2",
    "langgraph-supervisor>=0.0.29",
    "mcp>=1.14.1",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["agents"]
